Split automata generator between finite regexes and infinite regexes
The presence of Kleene operators inside simplified regular expressions
is used to determine whether its language is infinite or not. According
to this criterion, the regex-derived automata are divided, to allow for
the parallel lexing phase to distinguish between the two types. All the
finite-length regexes are grouped into one automata, while each of the
infinite-length regexes are converted into separate automata.